Comments

+++ python/ray/serve/api.py
@@ -1,4 +1,5 @@
 import atexit
+import time
 from functools import wraps
 import os
 from uuid import UUID
@@ -79,22 +80,22 @@ class Client:
         Shuts down all processes and deletes all state associated with the
         instance.
         """
-        if not self._shutdown:
+        if (not self._shutdown) and ray.is_initialized():
             ray.get(self._controller.shutdown.remote())
             ray.kill(self._controller, no_restart=True)
 
             # Wait for the named actor entry gets removed as well.
-            num_tries = 0
+            started = time.time()
             while True:
                 try:
                     ray.get_actor(self._controller_name)
-                    num_tries += 1
-                    if num_tries > 10:
-                        logger.info(
-                            "Waiting for controller actor be to removed "
-                            "from Ray actor registry. If this message "
-                            "occur too many times, please file a bug "
-                            "report.")
+                    if time.time() - started > 5:
+                        logger.warning(
+                            "Waited 5s for Serve to shutdown gracefully but "
+                            "the controller is still not cleaned up. "
+                            "You can ignore this warning if you are shutting "
+                            "down the Ray cluster.")
+                        break
                 except ValueError:  # actor name is removed
                     break
